+++ src/aap_eda/wsgi.py
@@ -25,7 +25,7 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-from aap_eda.logging.startup import startup_logging
+from aap_eda.utils.logging import startup_logging
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "aap_eda.settings.default")
 logger = logging.getLogger(__name__)
